requirements.txt
alembic==1.9.1
aniso8601==9.0.1
apispec==6.0.2
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
flask-apispec==0.11.4
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.2.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==22.0
passlib==1.7.4
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.11.0
